backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.2
boto3==1.42.49
botocore==1.42.49
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.5
dj-database-url==3.1.0
Django==6.0.2
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-camel-case==1.0.2
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
frozenlist==1.8.0
google-auth==2.48.0
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.5.4
maileroo==1.0.0
multidict==6.7.1
mutagen==1.47.0
openpyxl==3.1.5
orjson==3.11.5
packaging==25.0
phonenumbers==9.0.21
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
py-vapid==1.9.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
PyMySQL==1.1.2
PyPDF2==3.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==22.5
pywebpush==2.3.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
